Remove has_variable from templates
Without this patch the use of has_variable? in templates are causing
logic errors.  PXE configuration options are being added when they
should not be.

This patch replaces has_variable? with instance variables.  This fixes
the problem and correctly produces templates.

NOTE: The concat module assumes an empty string passed to content is the
same as content being undefined.  As a result, the template fragment
uses a static header comment to always make sure the string is not
empty.