Make the dhcp module more reusable
Without this patch large portions of the dhcpd.conf are solely under the
control of the dhcp module itself.  This is a problem for 3rd parties
like myself who want to change some of the behavior configured in this
file by these resources.

This patch fixes the problem in a non-breaking way by exposing all of
the configuration file fragments flowing into dhcpd.conf as class
parameters of the main dhcp class.  3rd parties may now override the
header using their own content like so:

    class { dhcp:
      dhcp_conf_header => template("site/my_dhcpd.conf.header")
    }

Conflicts:

	manifests/init.pp