Add support for iPXE chain loading
This commit adds support for iPXE chain loading, which allows a user to
specify a ipxe file and bootstrap script. When configured pxe will load
ipxe, which will then load the bootstrap script. This is helpful for
applications that use ipxe to load, like razor. A parameters section is
added to describe how to configure ipxe chainloading. Without this change it
is not possible to chain pxe to ipxe. Brief documentation is added for
the new functionality.

Additionally, this commit updates the handling of pxe and ipxe params to ensure that
the pxeserver and filename are both set, and ipxe bootstrap and filename
are both present when either is set. Without this change it is possible
to provide incomplete data to puppet and generate an invalid
configuration.

Lastly, this adds test cases to validate configuration placed by ipxe and
pxe parameters. Without this change the pxe configuration lacks test
coverage.